+++ Tests/MapboxSearchTests/Common/Models/Search Result/SearchResultSuggestionImplTests.swift
@@ -46,6 +46,48 @@ class SearchResultSuggestionImplTests: XCTestCase {
         XCTAssertEqual(suggestionImpl.namePreferred, coreResult.namePreferred)
     }
 
+    func testSuccessfulInitForCategories() throws {
+        let coreResult = CoreSearchResultStub(
+            id: "sample-2",
+            mapboxId: "sample-2",
+            type: .poi,
+            centerLocation: nil,
+            categories: ["category1", "category2"]
+        )
+        let suggestionImpl = try XCTUnwrap(SearchResultSuggestionImpl(
+            coreResult: coreResult,
+            response: CoreSearchResponseStub(
+                id: 42,
+                options: .sample1,
+                result: .success([])
+            )
+        ))
+        XCTAssertEqual(suggestionImpl.suggestionType, .POI)
+        XCTAssertEqual(suggestionImpl.name, coreResult.names.first)
+        XCTAssertEqual(suggestionImpl.categories, coreResult.categories)
+    }
+
+    func testSuccessfulInitForCategoryIds() throws {
+        let coreResult = CoreSearchResultStub(
+            id: "sample-2",
+            mapboxId: "sample-2",
+            type: .poi,
+            centerLocation: nil,
+            categoryIDs: ["categoryID1", "categoryID2"]
+        )
+        let suggestionImpl = try XCTUnwrap(SearchResultSuggestionImpl(
+            coreResult: coreResult,
+            response: CoreSearchResponseStub(
+                id: 42,
+                options: .sample1,
+                result: .success([])
+            )
+        ))
+        XCTAssertEqual(suggestionImpl.suggestionType, .POI)
+        XCTAssertEqual(suggestionImpl.name, coreResult.names.first)
+        XCTAssertEqual(suggestionImpl.categoryIds, coreResult.categoryIDs)
+    }
+
     func testFailedInit() throws {
         XCTAssertNil(SearchResultSuggestionImpl(
             coreResult: CoreSearchResultStub(