+++ Tests/MapboxSearchTests/Common/Stubs&Models/SearchResultStub.swift
@@ -7,6 +7,7 @@ class SearchResultStub: SearchResult {
         mapboxId: String?,
         accuracy: SearchResultAccuracy? = nil,
         categories: [String]? = nil,
+        categoryIds: [String]? = nil,
         name: String,
         matchingName: String?,
         serverIndex: Int?,
@@ -24,6 +25,7 @@ class SearchResultStub: SearchResult {
         self.mapboxId = mapboxId
         self.accuracy = accuracy
         self.categories = categories
+        self.categoryIds = categoryIds
         self.name = name
         self.matchingName = matchingName
         self.serverIndex = serverIndex
@@ -44,6 +46,7 @@ class SearchResultStub: SearchResult {
     var mapboxId: String?
     var accuracy: SearchResultAccuracy?
     var categories: [String]?
+    var categoryIds: [String]?
     var name: String
     var matchingName: String?
     var serverIndex: Int?