+++ Tests/MapboxSearchTests/Legacy/CodablePersistentServiceTests.swift
@@ -41,6 +41,7 @@ class CodablePersistentServiceTests: XCTestCase {
             serverIndex: nil,
             accuracy: nil,
             categories: [],
+            categoryIds: [],
             resultType: .address(subtypes: [.address]),
             searchRequest: .init(query: "Sample", proximity: nil)
         )